requirements.txt
flask==3.0.2
openai==1.12.0
requests>=2.25.0
python-dotenv>=0.15.0
flask-cors==4.0.0 